Restructure app further
