+++ test/index.test.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { DarkModeSwitch } from '../src';
+import { DarkModeSwitch, ThemeMode } from '../src';
 
 describe('it renders', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
     ReactDOM.render(
-      <DarkModeSwitch onChange={() => {}} checked={false} />,
+      <DarkModeSwitch onChange={() => {}} themeMode={ThemeMode.System} />,
       div
     );
     ReactDOM.unmountComponentAtNode(div);
